backend/py-result-getter-gsheets/requirements.txt
azure-core==1.26.3
azure-functions==1.13.3
azure-storage-blob==12.15.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
isodate==0.6.1
msal==1.21.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
requests==2.28.2
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Django==4.1.7